Домашки сделаны в отдельных классах
