+++ src/com/company/Homework3.java
@@ -1,0 +1,4 @@
+package com.company;
+
+public class Homework3 {
+}
